requirements_core.txt
# Core Web Framework
fastapi
uvicorn
python-multipart

# CrewAI (Keep exact versions)
crewai==0.130.0 
crewai-tools==0.47.1

# LangChain + Gemini (Google Generative AI)
langchain-google-genai
google-generativeai

# PDF Processing
pypdf2

# Environment Variables
python-dotenv
